Include all card game states in bins
Originally we were building round-length frequency bins starting from
the beginning of the round.  It is possible to build frequencies
starting from any point in the given round, i.e. after the first card,
or after the second card, etc.

This allows us to build additional plots that include more detail than
the original histogram.  For instance, we could build a heatmap of the
average round length after 2 gems and 2 hazards, which when clicked,
displays the histogram of all round lengths for 2 gems + 2 hazards.