Add page layout css
Page attempts to take up 100% width and at least 100% height.  On small
screens, plots are stacked.  On large screens, plots are side-by-side.

In order to get the plots to work right, they are taken out of flow
(position: absolute).  The resize handler (Plotly.plots.resize) checks
the parent element's size and attempts to fit it.  If the plots are in
flow, they will never get smaller.